requirements.txt
Flask
flask_sqlalchemy
werkzeug
psycopg2
flask-wtf
flask-login
gunicorn
flask-testing
coverage
coveralls
flask-wtf
flask-marshmallow
pymysql
flask-bootstrap 
flask-migrate
mashmallow-sqlalchemy
